+++ pytsite/auth_http_api/res/assets/js/pytsite-auth-http-api.js
@@ -1,6 +1,6 @@
 define(['pytsite-http-api'], function (httpApi) {
     function isAnonymous() {
-        return httpApi.get('auth/is_anonymous');
+        return httpApi.get('auth/is_anonymous', null, 2);
     }
 
     return {